Add Message component and request utils
